Move config derive and start on testing
