+++ Cargo.toml
@@ -3,6 +3,7 @@
 members = [
     "mirrord-auth",
     "mirrord-config",
+    "mirrord-config/derive",
     "mirrord-protocol",
     "mirrord-agent",
     "mirrord-layer",
